+++ Sources/Sentry/SentrySessionReplayIntegration.m
@@ -68,27 +68,43 @@ static SentryTouchTracker *_touchTracker;
 - (instancetype)initForManualUse:(nonnull SentryOptions *)options
 {
     if (self = [super init]) {
-        [self setupWith:options.sessionReplay enableTouchTracker:options.enableSwizzling];
+        [self setupWith:options.sessionReplay
+                    enableTouchTracker:options.enableSwizzling
+            enableExperimentalRenderer:options.sessionReplay.enableExperimentalViewRenderer
+               enableFastViewRendering:options.sessionReplay.enableFastViewRendering];
         [self startWithOptions:options.sessionReplay fullSession:YES];
     }
     return self;
 }
 
 - (BOOL)installWithOptions:(nonnull SentryOptions *)options
+    enableFastViewRendering:(BOOL)enableFastViewRendering
 {
     if ([super installWithOptions:options] == NO) {
         return NO;
     }
 
-    [self setupWith:options.sessionReplay enableTouchTracker:options.enableSwizzling];
+    [self setupWith:options.sessionReplay
+                enableTouchTracker:options.enableSwizzling
+        enableExperimentalRenderer:options.sessionReplay.enableExperimentalViewRenderer
+           enableFastViewRendering:options.sessionReplay.enableFastViewRendering];
     return YES;
 }
 
-- (void)setupWith:(SentryReplayOptions *)replayOptions enableTouchTracker:(BOOL)touchTracker
+- (void)setupWith:(SentryReplayOptions *)replayOptions
+            enableTouchTracker:(BOOL)touchTracker
+    enableExperimentalRenderer:(BOOL)enableExperimentalRenderer
+       enableFastViewRendering:(BOOL)enableFastViewRendering
 {
     _replayOptions = replayOptions;
     _rateLimits = SentryDependencyContainer.sharedInstance.rateLimits;
-    id<SentryViewRenderer> viewRenderer = [[SentryDefaultViewRenderer alloc] init];
+    id<SentryViewRenderer> viewRenderer;
+    if (enableExperimentalRenderer) {
+        viewRenderer = [[SentryExperimentalViewRenderer alloc]
+            initWithEnableFastViewRendering:enableFastViewRendering];
+    } else {
+        viewRenderer = [[SentryDefaultViewRenderer alloc] init];
+    }
     _viewPhotographer = [[SentryViewPhotographer alloc] initWithRenderer:viewRenderer
                                                            redactOptions:replayOptions];