+++ Tests/SentryTests/Helper/SentryEnabledFeaturesBuilderTests.swift
@@ -2,36 +2,43 @@
 import XCTest
 
 final class SentryEnabledFeaturesBuilderTests: XCTestCase {
-
+    
     func testDefaultFeatures() throws {
-        let features = SentryEnabledFeaturesBuilder.getEnabledFeatures(options: Options())
+        // -- Arrange --
+        let options = Options()
         
+        // -- Act --
+        let features = SentryEnabledFeaturesBuilder.getEnabledFeatures(options: options)
+        
+        // -- Assert --
         XCTAssertEqual(features, ["captureFailedRequests"])
     }
     
     func testEnableAllFeatures() throws {
-        
+        // -- Arrange --
         let options = Options()
         options.enablePerformanceV2 = true
         options.enableTimeToFullDisplayTracing = true
         options.swiftAsyncStacktraces = true
-
+        
 #if os(iOS) || os(macOS) || targetEnvironment(macCatalyst)
         options.enableAppLaunchProfiling = true
 #endif // os(iOS) || os(macOS) || targetEnvironment(macCatalyst)
-
+        
 #if os(iOS) || os(tvOS)
 #if canImport(UIKit) && !SENTRY_NO_UIKIT
         options.enablePreWarmedAppStartTracing = true
 #endif // canImport(UIKit)
 #endif // os(iOS) || os(tvOS)
-      
+        
 #if os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
         options.enableAppHangTrackingV2 = true
 #endif //os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
         
+        // -- Act --
         let features = SentryEnabledFeaturesBuilder.getEnabledFeatures(options: options)
         
+        // -- Assert --
         XCTAssert(features.contains("captureFailedRequests"))
         XCTAssert(features.contains("performanceV2"))
         XCTAssert(features.contains("timeToFullDisplayTracing"))
@@ -40,7 +47,7 @@ final class SentryEnabledFeaturesBuilderTests: XCTestCase {
 #if os(iOS) || os(macOS) || targetEnvironment(macCatalyst)
         XCTAssert(features.contains("appLaunchProfiling"))
 #endif // os(iOS) || os(macOS) || targetEnvironment(macCatalyst)
-    
+        
 #if os(iOS) || os(tvOS)
 #if canImport(UIKit) && !SENTRY_NO_UIKIT
         XCTAssert(features.contains("preWarmedAppStartTracing"))
@@ -50,16 +57,52 @@ final class SentryEnabledFeaturesBuilderTests: XCTestCase {
 #if os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
         XCTAssert(features.contains("appHangTrackingV2"))
 #endif //os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
-        
     }
     
     func testEnablePersistingTracesWhenCrashing() {
+        // -- Arrange --
         let options = Options()
         
         options.enablePersistingTracesWhenCrashing = true
         
+        // -- Act --
         let features = SentryEnabledFeaturesBuilder.getEnabledFeatures(options: options)
         
+        // -- Assert --
         XCTAssert(features.contains("persistingTracesWhenCrashing"))
+    }
+    
+    func testEnableExperimentalViewRenderer_isEnabled_shouldAddFeature() throws {
+#if SENTRY_TARGET_REPLAY_SUPPORTED
+        // -- Arrange --
+        let options = Options()
+        
+        options.sessionReplay.enableExperimentalViewRenderer = true
+        
+        // -- Act --
+        let features = SentryEnabledFeaturesBuilder.getEnabledFeatures(options: options)
+        
+        // -- Assert --
+        XCTAssert(features.contains("experimentalViewRenderer"))
+#else
+        throw XCTSkip("Feature not available on this platform.")
+#endif
+    }
+    
+    func testEnableFastViewRendering_isEnabled_shouldAddFeature() throws {
+#if SENTRY_TARGET_REPLAY_SUPPORTED
+        // -- Arrange --
+        let options = Options()
+        
+        options.sessionReplay.enableFastViewRendering = true
+        
+        // -- Act --
+        let features = SentryEnabledFeaturesBuilder.getEnabledFeatures(options: options)
+        
+        // -- Assert --
+        XCTAssert(features.contains("fastViewRenderer"))
+#else
+        throw XCTSkip("Feature not available on this platform.")
+#endif
     }
 }